requirements.txt
alembic
aniso8601
blinker
click
Flask==1.0.2
flask-cors
flask-feide-gatekeeper==0.0.6
flask-fixtures
flask-graphql
flask-migrate
flask-shell-ipython
flask-sqlalchemy
graphene==2.1.3
graphene-file-upload==1.2.1
graphene-sqlalchemy
gunicorn
psycopg2-binary
pynacl
python-json-logger
pytz
setuptools
sqlalchemy==1.2.15
sqlalchemy-continuum
sqlalchemy-json
sqlalchemy-utils
werkzeug

# Our own dependencies
sap-client==0.2.0
